Move reactor library instrumentation from instrumentation-core to instrumentation
