--- instrumentation-core/reactor-3.1/reactor-3.1.gradle
@@ -1,28 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply from: "$rootDir/gradle/instrumentation-library.gradle"
-
-// need to override archivesBaseName set in instrumentation-library.gradle
-archivesBaseName = "reactor-3.1"
-
-dependencies {
-  library group: 'io.projectreactor', name: 'reactor-core', version: '3.1.0.RELEASE'
-
-  latestDepTestLibrary group: 'io.projectreactor', name: 'reactor-core', version: '3.+'
-  // Looks like later versions on reactor need this dependency for some reason even though it is marked as optional.
-  latestDepTestLibrary group: 'io.micrometer', name: 'micrometer-core', version: '1.+'
-}
