+++ settings.gradle
@@ -190,6 +190,8 @@ include ':instrumentation:play-ws:play-ws-testing'
 include ':instrumentation:rabbitmq-2.7:javaagent'
 include ':instrumentation:ratpack-1.4:javaagent'
 include ':instrumentation:reactor-3.1:javaagent'
+include ':instrumentation:reactor-3.1:library'
+include ':instrumentation:reactor-3.1:testing'
 include ':instrumentation:reactor-netty-0.9:javaagent'
 include ':instrumentation:reactor-netty-1.0:javaagent'
 include ':instrumentation:rediscala-1.8:javaagent'
@@ -230,7 +232,6 @@ include ':instrumentation:vertx-web-3.0'
 include ':instrumentation:vertx-reactive-3.5:javaagent'
 include ':instrumentation:wicket-8.0:javaagent'
 
-include ':instrumentation-core:reactor-3.1'
 include ':instrumentation-core:servlet-2.2'
 
 // benchmark
